Core/PacketIO: Updated most packet structures to 9.0.1
Port From (https://github.com/TrinityCore/TrinityCore/commit/cab4c87d2d7f6d734ef067d6bf50f4b1d338a7bc)

+++ Source/Game/Networking/Packets/ChannelPackets.cs
@@ -115,7 +115,7 @@ namespace Game.Networking.Packets
             _worldPacket.WriteBits(ChannelWelcomeMsg.GetByteCount(), 11);
             _worldPacket.WriteUInt32((uint)ChannelFlags);
             _worldPacket.WriteInt32(ChatChannelID);
-            _worldPacket.WriteUInt64((ulong)InstanceID);
+            _worldPacket.WriteUInt64(InstanceID);
             _worldPacket.WritePackedGuid(ChannelGUID);
             _worldPacket.WriteString(Channel);
             _worldPacket.WriteString(ChannelWelcomeMsg);
@@ -123,7 +123,7 @@ namespace Game.Networking.Packets
 
         public string ChannelWelcomeMsg = "";
         public int ChatChannelID;
-        public int InstanceID;
+        public ulong InstanceID;
         public ChannelFlags ChannelFlags;
         public string Channel = "";
         public ObjectGuid ChannelGUID;
